+++ Source/Core/Core/WiiUtils.cpp
@@ -2,7 +2,8 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
-#include "UICommon/WiiUtils.h"
+#include "Core/WiiUtils.h"
+
 #include "Common/CommonTypes.h"
 #include "Common/MsgHandler.h"
 #include "Core/ConfigManager.h"
