+++ Source/Core/Core/WiiUtils.h
